Added some stream classes

+++ src/SystemLayer/Windows/SystemLayerWindows_IO.cpp
@@ -93,11 +93,11 @@ namespace nl::systemlayer::defaults
         return true;
     }
 
-    static size_t Read(FileHandle fp, void* ptr, size_t count)
+    static int64_t Read(FileHandle fp, void* ptr, int64_t count)
     {
         HANDLE hFile = (HANDLE)fp;
         
-        size_t remaining = count;
+        int64_t remaining = count;
         while (remaining != 0)
         {
             DWORD dw;
@@ -112,11 +112,11 @@ namespace nl::systemlayer::defaults
         return count - remaining;
     }
 
-    static size_t Write(FileHandle fp, const void* ptr, size_t count)
+    static int64_t Write(FileHandle fp, const void* ptr, int64_t count)
     {
         HANDLE hFile = (HANDLE)fp;
         
-        size_t remaining = count;
+        int64_t remaining = count;
         while (remaining != 0)
         {
             DWORD dw;
